End of 330

+++ petsoft/src/actions/actions.ts
@@ -6,13 +6,17 @@ import prisma from '@/lib/db';
 import { delay } from '@/lib/utils';
 import { Pet } from '@prisma/client';
 import { PetEssentials } from '@/lib/types';
+import { petFormSchema, petIdSchema } from '@/lib/validations';
 
-export async function addNewPet(pet: PetEssentials) {
+export async function addNewPet(pet: unknown) {
   await delay(1000);
+
+  const validatedPet = petFormSchema.safeParse(pet);
+  if (!validatedPet.success) return { message: 'Invalid pet data' };
 
   try {
     await prisma.pet.create({
-      data: pet,
+      data: validatedPet.data,
     });
   } catch (error) {
     return { message: 'Failed to add pet' };
@@ -21,15 +25,20 @@ export async function addNewPet(pet: PetEssentials) {
   revalidatePath('/app', 'layout');
 }
 
-export async function editPet(petId: Pet['id'], newPetData: PetEssentials) {
+export async function editPet(petId: unknown, newPetData: unknown) {
   await delay(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPetData);
+  if (!validatedPetId.success || !validatedPet.success)
+    return { message: 'Invalid pet data' };
 
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPetData,
+      data: validatedPet.data,
     });
   } catch (error) {
     return { message: 'Failed to edit pet' };
@@ -38,13 +47,16 @@ export async function editPet(petId: Pet['id'], newPetData: PetEssentials) {
   revalidatePath('/app', 'layout');
 }
 
-export async function deletePet(petId: Pet['id']) {
+export async function deletePet(petId: unknown) {
   await delay(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  if (!validatedPetId.success) return { message: 'Invalid pet data' };
 
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
